Break up conftest
