+++ kasa/tests/test_feature_brightness
@@ -1,0 +1,12 @@
+from kasa.smart import SmartDevice
+
+from .conftest import (
+    brightness_smart,
+)
+
+
+@brightness_smart
+async def test_brightness_component(dev: SmartDevice):
+    """Placeholder to test framwework component filter."""
+    assert isinstance(dev, SmartDevice)
+    assert "brightness" in dev._components
